Reconcile pipeline with the original repo (reuse Trainer; fix repo bugs)
Engage with the existing code instead of running beside it.

Fix original-repo bugs that prevented it from running (incl. on CPU):
- config.py: add the missing Config.set_args() (test.py/demo.py called it ->
  crash) with graceful CPU fallback (use_cuda, num_gpus).
- model.py: soft_argmax built torch.arange with no device -> GPU crash; now on
  heatmaps.device (runs on CPU and GPU).
- base.py: Tester/Transformer used unconditional .cuda(); add CPU guards,
  map_location, and module.-prefix matching for cross-device checkpoints.
- train.py: implement the --gpu/--backbone/--continue args the README documents
  but that were never parsed.
- dataset.py: float [0,255] patch fed to ToTensor (which only /255 for uint8)
  -> wrong Normalize stats; cast to uint8 first.
Verified: `python main/train.py --backbone LPSKI` now trains on CPU (Dummy).

Integrate the pipeline into the original training machinery:
- data/CrawlPipeline/CrawlPipeline.py: a 2D COCO loader (modeled on MSCOCO)
  that reads the pipeline's crawled/annotated outputs and produces the repo's
  canonical 19-joint, depth-less target; optionally oversamples by the Stage-4
  embedding weights (rebalancing with no change to base.py).
- stage05_train.py: drop the parallel training loop; configure cfg and run the
  ORIGINAL base.Trainer on CrawlPipeline. Writes the repo-format snapshot
  (loadable by main/test.py) plus a backbone checkpoint for demo/ONNX.
- common.py/inference.py/export_onnx.py/web: adopt the repo 19-joint convention
  so a Trainer-trained model is consumed unchanged (raw stage-3 annotations stay
  COCO-17).

Docs + tests:
- pipeline/INTEGRATION.md: original architecture, bugs fixed, how it connects.
- tests: assert stage 5 trains via the repo Trainer and writes a repo snapshot.
Verified: 8/8 smoke tests pass on a fresh crawl.

+++ data/dataset.py
@@ -100,8 +100,10 @@ class DatasetLoader(Dataset):
         joint_img[:, 2] = joint_img[:, 2] * cfg.depth_dim
         
         if self.is_train:
-            img_patch = self.transform(img_patch)
-            
+            # cast to uint8 so torchvision ToTensor divides by 255 before
+            # Normalize (it does NOT rescale float arrays -> wrong stats).
+            img_patch = self.transform(img_patch.astype(np.uint8))
+
             if self.ref_joints_name is not None:
                 joint_img = transform_joint_to_other_db(joint_img, self.joints_name, self.ref_joints_name) 
                 joint_vis = transform_joint_to_other_db(joint_vis, self.joints_name, self.ref_joints_name)
@@ -112,7 +114,7 @@ class DatasetLoader(Dataset):
 
             return img_patch, joint_img, joint_vis, joints_have_depth
         else:
-            img_patch = self.transform(img_patch)
+            img_patch = self.transform(img_patch.astype(np.uint8))
             return img_patch
 
     def __len__(self):
